Move code around.
