--- crates/notifier/src/subscribers/mod.rs
@@ -1,3 +1,0 @@
-mod webhooks;
-
-pub use webhooks::WebhooksSubscriber;
